requirements.txt
asgiref==3.5.2
Django==4.0.6
django-autoslug-iplweb==1.9.4.dev0
django-countries==6.1.2
django-extensions==3.2.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
Pillow==9.2.0
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
whitenoise==5.1.0
